part3/flin-api/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
boto3==1.36.24
botocore==1.36.24
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-database-url==2.3.0
Django==5.1.6
django-admin-rangefilter==0.13.2
django-cors-headers==4.7.0
django-filter==25.1
django-json-widget==2.0.1
django-safedelete==1.4.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
drf-yasg==1.21.8
google-auth==2.38.0
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
jmespath==1.0.1
Markdown==3.7
packaging==24.2
phonenumbers==8.13.55
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rsa==4.9
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
